back stab and spell system added

+++ Assets/Scripts/Ui/StaminaBar.cs
@@ -7,7 +7,7 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxStamina(int maxStamina)
+    public void SetMaxStamina(float maxStamina)
     {
         slider.maxValue = maxStamina;
         slider.value = maxStamina;
